Change pages and add Home page
