+++ frontend/src/App.js
@@ -1,9 +1,10 @@
 import React from "react";
 import { Switch, Router, Route, Redirect } from "react-router";
 import { createBrowserHistory } from "history";
-import Home from "./pages/Home/Home";
 import Header from "./components/header/header";
 import * as toastr from "toastr";
+import Home from './pages/Home/Home';
+import Presentation from "./pages/Presentation/Presentation";
 
 const App = () => {
     const history = createBrowserHistory();
@@ -13,8 +14,10 @@ const App = () => {
         <Header />
         <Router history={history}>
             <Switch>
+                <Route path="/presentation" component={Presentation} />
                 <Route path="/home" component={Home} />
-                <Redirect to="/home" />
+                
+                <Redirect to="/presentation" />
             </Switch>
         </Router>
       </div>
