+++ frontend/src/pages/Home/Home.jsx
@@ -1,345 +1,35 @@
-import React from "react";
-import Page from "../../components/page/page";
+import classNames from "classnames";
 import Button from "../../components/button/button";
-import InputField from "../../components/input-field/inputField";
-import CheckerField from "../../components/checker-field/checkerField";
-import ImageSelectField from "../../components/image-select-field/imageSelectField";
-import SelectField from "../../components/select-field/selectField";
-import * as toastr from "toastr";
-import RangeField from "../../components/range-field/rangeField";
-
+import Page from "../../components/page/page";
+import classes from "./Home.module.scss";
+import bannerImage from "../../assets/banner-image.svg"
 
 const Home = (props) => {
-    function testNotifications() {
-        toastr.warning('Lorem ipsum dolor sit amet, consectetur', 'Lorem ipsum');
-        toastr.success('Lorem ipsum dolor sit amet, consectetur', 'Lorem ipsum');
-        toastr.error('Lorem ipsum dolor sit amet, consectetur', 'Lorem ipsum');
-    }
-    return (
-        <Page title="Bemutató oldal">
-            <h3>Gombok</h3>
+   return (
+      <Page noCard>
+         <div className={classNames(classes.banner)}>
+            <div className={classNames(classes.row)}>
+               <div className={classNames(classes.texts)}>
+                  <h1>Tervezd meg a saját Slack Botod</h1>
+                  <p>
+                     Donec ipsum tortor, bibendum fringilla mauris at, mattis mattis metus. 
+                     Mauris bibendum, mauris id scelerisque laoreet, dui tortor lacinia felis, vitae porta.
+                  </p>
+               </div>
+               <img className={classNames(classes.bannerImage)} src={bannerImage} />
+            </div>            
             <div>
-                <Button>Default</Button>
-                <hr/>
-                <Button color="blue">blue</Button>
-                <Button color="blue" isHollow={true}>blue isHollow</Button>
-                <Button color="blue" isLarge={true}>blue isLarge</Button>
-                <Button color="blue" isHollow={true} isLarge={true}>blue isHollow isLarge</Button>
-                <hr/>
-                <Button color="red">red</Button>
-                <Button color="red" isHollow={true}>red isHollow</Button>
-                <Button color="red" isLarge={true}>red isLarge</Button>
-                <Button color="red" isHollow={true} isLarge={true}>red isHollow isLarge</Button>
-                <hr/>
-                <Button color="yellow">yellow</Button>
-                <Button color="yellow" isHollow={true}>yellow isHollow</Button>
-                <Button color="yellow" isLarge={true}>yellow isLarge</Button>
-                <Button color="yellow" isHollow={true} isLarge={true}>yellow isHollow isLarge</Button>
-                <hr/>
-                <Button onClick={() => alert("Test alert")}>onClick alert</Button>
-            </div>
-            <h3>Sima inputok</h3>
-            <div>
-                <InputField label="Teszt label"/>
-                <InputField label="Teszt label number" type="number"/>
-                <InputField label="Teszt label password" type="password"/>
-                <InputField label="Teszt label password" type="password"/>
-                <InputField label="Teszt label with error" type="password" reduxFormData={{
-                    input: {},
-                    meta: {
-                        touched: true,
-                        error: "Ez egy error, ajjaj!"
-                    },
-                }}/>
-                <h4>Checkboxes</h4>
-                <CheckerField label="Teszt label normal" options={[
-                    {
-                        label: "Option 1",
-                        value: "option_1"
-                    },
-                    {
-                        label: "Option 2",
-                        value: "option_2"
-                    },
-                    {
-                        label: "Option 3",
-                        value: "option_3"
-                    }
-                ]}/>
-                <CheckerField label="Teszt label red" color="red" options={[
-                    {
-                        label: "Option 1",
-                        value: "option_1"
-                    },
-                    {
-                        label: "Option 2",
-                        value: "option_2"
-                    },
-                    {
-                        label: "Option 3",
-                        value: "option_3"
-                    }
-                ]}/>
-                <CheckerField label="Teszt label red" color="yellow" options={[
-                    {
-                        label: "Option 1",
-                        value: "option_1"
-                    },
-                    {
-                        label: "Option 2",
-                        value: "option_2"
-                    },
-                    {
-                        label: "Option 3",
-                        value: "option_3"
-                    }
-                ]}/>
-                <CheckerField label="Teszt label red" reduxFormData={{
-                    input: {},
-                    meta: {
-                        touched: true,
-                        error: "Ez egy error, ajjaj!"
-                    },
-                }} options={[
-                    {
-                        label: "Option 1",
-                        value: "option_1"
-                    },
-                    {
-                        label: "Option 2",
-                        value: "option_2"
-                    },
-                    {
-                        label: "Option 3",
-                        value: "option_3"
-                    }
-                ]}/>
-                <h4>Radio</h4>
-                <CheckerField type="radio" label="Teszt label normal" options={[
-                    {
-                        label: "Option 1",
-                        value: "option_1"
-                    },
-                    {
-                        label: "Option 2",
-                        value: "option_2"
-                    },
-                    {
-                        label: "Option 3",
-                        value: "option_3"
-                    }
-                ]}/>
-                <CheckerField type="radio" label="Teszt label red" color="red" options={[
-                    {
-                        label: "Option 1",
-                        value: "option_1"
-                    },
-                    {
-                        label: "Option 2",
-                        value: "option_2"
-                    },
-                    {
-                        label: "Option 3",
-                        value: "option_3"
-                    }
-                ]}/>
-                <CheckerField type="radio" label="Teszt label yellow" color="yellow" options={[
-                    {
-                        label: "Option 1",
-                        value: "option_1"
-                    },
-                    {
-                        label: "Option 2",
-                        value: "option_2"
-                    },
-                    {
-                        label: "Option 3",
-                        value: "option_3"
-                    }
-                ]}/>
-                <CheckerField type="radio" label="Teszt label error" reduxFormData={{
-                    input: {},
-                    meta: {
-                        touched: true,
-                        error: "Ez egy error, ajjaj!"
-                    },
-                }} options={[
-                    {
-                        label: "Option 1",
-                        value: "option_1"
-                    },
-                    {
-                        label: "Option 2",
-                        value: "option_2"
-                    },
-                    {
-                        label: "Option 3",
-                        value: "option_3"
-                    }
-                ]}/>
-                <h2>Image select fields</h2>
-                <ImageSelectField label="Teszt label" reduxFormData={{
-                    input: {},
-                    meta: {
-                        touched: false,
-                        error: "Ez egy error, ajjaj!"
-                    },
-                }} options={[
-                    {
-                        image: "https://placekitten.com/300/300",
-                        value: "option_1"
-                    },
-                    {
-                        image: "https://placekitten.com/400/400",
-                        value: "option_2"
-                    },
-                    {
-                        image: "https://placekitten.com/300/300",
-                        value: "option_1"
-                    },
-                    {
-                        image: "https://placekitten.com/400/400",
-                        value: "option_2"
-                    },
-                    {
-                        image: "https://placekitten.com/300/300",
-                        value: "option_1"
-                    },
-                    {
-                        image: "https://placekitten.com/400/400",
-                        value: "option_2"
-                    },
-                    {
-                        image: "https://placekitten.com/500/400",
-                        value: "option_3"
-                    }
-                ]}/>
-                <ImageSelectField label="Teszt label error" reduxFormData={{
-                    input: {},
-                    meta: {
-                        touched: true,
-                        error: "Ez egy error, ajjaj!"
-                    },
-                }} options={[
-                    {
-                        image: "https://placekitten.com/300/300",
-                        value: "option_1"
-                    },
-                    {
-                        image: "https://placekitten.com/400/400",
-                        value: "option_2"
-                    },
-                    {
-                        image: "https://placekitten.com/300/300",
-                        value: "option_1"
-                    },
-                    {
-                        image: "https://placekitten.com/400/400",
-                        value: "option_2"
-                    },
-                    {
-                        image: "https://placekitten.com/300/300",
-                        value: "option_1"
-                    },
-                    {
-                        image: "https://placekitten.com/400/400",
-                        value: "option_2"
-                    },
-                    {
-                        image: "https://placekitten.com/500/400",
-                        value: "option_3"
-                    }
-                ]}/>
-                <h2>Select fields</h2>
-                <SelectField label="Teszt label" reduxFormData={{
-                    input: {},
-                    meta: {
-                        touched: false,
-                        error: "Ez egy error, ajjaj!"
-                    },
-                }} options={[
-                    {
-                        label: "Please select...",
-                        value: "option_1"
-                    },
-                    {
-                        label: "Lorem ipsum dolor sit amet",
-                        value: "option_2"
-                    },
-                    {
-                        label: "Lorem ipsum dolor sit amet",
-                        value: "option_1"
-                    },
-                    {
-                        label: "Lorem ipsum dolor sit amet",
-                        value: "option_2"
-                    },
-                    {
-                        label: "Lorem ipsum dolor sit amet",
-                        value: "option_1"
-                    },
-                    {
-                        label: "Lorem ipsum dolor sit amet",
-                        value: "option_2"
-                    },
-                    {
-                        label: "Lorem ipsum dolor sit amet",
-                        value: "option_3"
-                    }
-                ]}/>
-                <SelectField label="Teszt label error" reduxFormData={{
-                    input: {},
-                    meta: {
-                        touched: true,
-                        error: "Ez egy error, ajjaj!"
-                    },
-                }} options={[
-                    {
-                        label: "Please select...",
-                        value: "option_1"
-                    },
-                    {
-                        label: "Lorem ipsum dolor sit amet",
-                        value: "option_2"
-                    },
-                    {
-                        label: "Lorem ipsum dolor sit amet",
-                        value: "option_1"
-                    },
-                    {
-                        label: "Lorem ipsum dolor sit amet",
-                        value: "option_2"
-                    },
-                    {
-                        label: "Lorem ipsum dolor sit amet",
-                        value: "option_1"
-                    },
-                    {
-                        label: "Lorem ipsum dolor sit amet",
-                        value: "option_2"
-                    },
-                    {
-                        label: "Lorem ipsum dolor sit amet",
-                        value: "option_3"
-                    }
-                ]}/>
-
-            </div>
-            <h2>toastr.js</h2>
-            <div>
-                <Button onClick={testNotifications}>click to display test notification</Button>
-            </div>
-            <h2>Range Slider</h2>
-            <div>
-                <RangeField label="range" />
-                <RangeField label="range min={0} max={500}" min={0} max={500}/>
-                <RangeField label="range min={-100} max={100}" min={-100} max={100}/>
-                <RangeField label="range min={-100} max={100} step={10}" min={-100} max={100} step={10}/>
-                <RangeField label="range yellow" color="yellow" />
-                <RangeField label="range red" color="red" />
-            </div>
-        </Page>
-    );
+               <Button isLarge>Regisztráció</Button>
+               <Button isLarge isHollow>Bejelentkezés</Button>
+            </div>            
+         </div>
+         <h2>Hogyan működik?</h2>
+         <div className={classNames(classes.twoColumn)}>
+            <p>Mauris bibendum, mauris id scelerisque laoreet, dui tortor lacinia felis, vitae porta justo dolor sit amet ipsum. Suspendisse ullamcorper blandit imperdiet. Curabitur maximus convallis ex sed ornare. Mauris vel dui id eros lobortis tempus at sit amet arcu. Nulla quis rhoncus augue. Aliquam consectetur erat eget leo ullamcorper, ac blandit ex vehicula. Praesent varius, justo vitae rhoncus pulvinar, nibh augue iaculis metus.</p>
+            <p>Mauris id scelerisque laoreet, dui tortor lacinia felis, vitae porta justo dolor sit amet ipsum. Suspendisse ullamcorper blandit imperdiet. Curabitur maximus convallis ex sed ornare. Mauris vel dui id eros lobortis tempus at sit amet arcu. Nulla quis rhoncus augue. Aliquam consectetur erat eget leo ullamcorper, ac blandit ex vehicula. Praesent varius, justo vitae rhoncus pulvinar, nibh augue iaculis metus, in tempor dui eros id diam.</p>
+         </div>
+      </Page>
+   );
 };
 
 export default Home;
