+++ frontend/src/styles/_defaults.scss
@@ -10,6 +10,8 @@
   -moz-osx-font-smoothing: grayscale;
 
   font-size: 10px;
+  color: $even-darker-blue;
+
   @include sm{
     font-size: 12px;
   }
@@ -38,4 +40,8 @@ button {
   all: unset;
   color: black;
   cursor: pointer;
+}
+
+p {
+  line-height: 1.5rem;
 }
