+++ frontend/src/styles/_variables.scss
@@ -6,7 +6,11 @@ $dark-blue: #296EB4;
 $dark-yellow: #B1740F;
 $dark-red: #981717;
 
+$even-darker-blue: #1B2C3E;
+
 $grey: #EFEFEF;
+
+$light-blue-gradient: linear-gradient(#97C9FB, #F3F8FE);
 
 $small: 1rem;
 $large: 1.25rem;
